+++ client/src/model/AppContext.ts
@@ -2,8 +2,8 @@ import React from 'react';
 import { IAppState, DefaultAppState } from './AppState';
 
 export const DefaultAppContextValue = {
-    state: DefaultAppState,
-    setState: (value: IAppState) => {},
+    appState: DefaultAppState,
+    setAppState: (value: IAppState) => {},
 };
 
 export const AppContext = React.createContext(DefaultAppContextValue);