+++ client/src/pages/_index.ts
@@ -1,4 +1,0 @@
-import SignInSide from './sign-up';
-import Assignment from './assignment';
-
-export { Assignment, SignInSide };