+++ server/api/_index.ts
@@ -1,0 +1,3 @@
+import * as UsersApi from './users/users.get';
+
+export default UsersApi;