+++ server/api/users/users.get.ts
@@ -1,12 +1,5 @@
 import { Request, Response } from 'express';
 
-const list = (req: Request, res: Response) => {
-    return new List<User> {
-        new User {
-            fname: "John",
-            lname: "Gilliland",
-            uname: 'jgilliland',
-            
-        }
-    }
-}
+export const list = (req: Request, res: Response) => {
+    return [];
+};