+++ server/database/migrations/20200914120000-user.ts
@@ -1,2 +1,0 @@
-import { QueryInterface, DataTypes } from 'sequelize';
-import { User } from '../models';