+++ server/database/models/_index.ts
@@ -1,7 +1,17 @@
 import { Sequelize } from 'sequelize';
-import { Config } from '../../config/_index';
+import Config from '../../config/_index';
 
-const dbConfig = Config[process.env.NODE_ENV];
-const sequelize = new Sequelize(dbConfig['connection'], dbConfig);
+var currentEnv = process.env.NODE_ENV;
+const dbConfig = Config[currentEnv];
+let sequelize: Sequelize;
+
+if (process.env.NODE_ENV === 'production') {
+    let sequelize = new Sequelize(); //(dbConfig);
+} else {
+    let sequelize = new Sequelize('database', 'user', 'password', {
+        dialect: 'sqlite',
+        storage: '../database/santa.db',
+    });
+}
 
 export default sequelize;